Move relevant enums and exceptiosn to music player
These should belong there since they're music player related. The
musicbox module should import these to make use of them.
